VectorMath cleanup

+++ src/Shared/GpxLayer.swift
@@ -798,7 +798,7 @@ class GpxLayer: CALayer, GetDiskCacheSize {
             
             if track.shapePaths[scale] == nil {
                 let epsilon = pow(Double(10.0), Double(scale)) / 256.0
-				track.shapePaths[scale] = track.shapePaths[0]?.pathWithReducePoints( CGFloat(epsilon) )
+				track.shapePaths[scale] = track.shapePaths[0]?.pathWithReducePoints( epsilon )
 			}
             //		DLog(@"reduce %ld to %ld\n",CGPathPointCount(track->shapePaths[0]),CGPathPointCount(track->shapePaths[scale]));
             layer.path = track.shapePaths[scale]
